+++ src/ui/components/splash-screen.js
@@ -1,26 +1,7 @@
 import {html, Component} from '../lib.js'
-import gsap from '../animations.js'
 import {getRuns} from '../../game/backend.js'
-
-function formatDate(timestamp) {
-	return new Intl.DateTimeFormat('en', {
-		dateStyle: 'medium',
-		// month: 'short',
-		// timeStyle: 'short',
-		hour12: false,
-	}).format(new Date(timestamp))
-}
-
-function timeSince(timestamp) {
-	const seconds = Math.floor((Date.now() - timestamp) / 1000)
-	if (seconds < 60) return 'just now'
-	if (seconds < 120) return 'a minute ago'
-	if (seconds < 3600) return `${Math.floor(seconds / 60)} minutes ago`
-	if (seconds < 7200) return 'an hour ago'
-	if (seconds < 86400) return `${Math.floor(seconds / 3600)} hours ago`
-	if (seconds < 172800) return 'yesterday'
-	return `${Math.floor(seconds / 86400)} days ago`
-}
+import {timeSince} from '../../utils.js'
+import gsap from '../animations.js'
 
 export default class SplashScreen extends Component {
 	constructor() {
@@ -59,15 +40,14 @@ export default class SplashScreen extends Component {
 						<li><a class="Button" href="/collection">Collection</a></li>
 						<li>
 							<a class="Button" href="/stats">Highscores</a>
-							${this.state.runs.length > 0 ? html`
-							<a class="LastRun" href=${`/stats/run?id=${this.state.runs[0].id}`}>
-								${timeSince(this.state.runs[0].createdAt)} someone ${this.state.runs[0].won ? 'won' : 'lost'}
-							</a>` : ''}
+							${this.state.runs.length > 0
+								? html` <a class="LastRun" href=${`/stats/run?id=${run.id}`}>
+										${timeSince(run.createdAt)} someone ${run.won ? 'won' : 'lost'}
+									</a>`
+								: ''}
 						</li>
 					</ul>
-					<p center>
-						<a href="/changelog">Changelog</a> & <a href="/manual">Manual</a>
-					</p>
+					<p center><a href="/changelog">Changelog</a> & <a href="/manual">Manual</a></p>
 				</div>
 			</article>
 		`
